centralising error handling more
