+++ src/auth-service/utils/common/index.js
@@ -21,7 +21,6 @@ const msgs = require("./email.msgs.util");
 const emailTemplates = require("./email.templates.util");
 const generateFilter = require("./generate-filter.util");
 const winstonLogger = require("./log-winston.util");
-const handleResponse = require("./response-handler.util");
 const slugUtils = require("./slug.util");
 const { deduplicator } = require("./slack-dedup-utility");
 
@@ -49,5 +48,4 @@ module.exports = {
   msgs,
   emailTemplates,
   generateFilter,
-  handleResponse,
 };
